move raw_fact parsing to objects.py
Made ``raw_fact`` to ``hamsterlib.Fact`` conversion a classmethod of
``Fact``. After long deliberation this still seems to be the cleanest
and consistent design.
It is worth pointing out that this is realy just about parsing and
converting/encapsulating information stored in the ``raw_fact``.
At this point no assumption can and should be made about how any of it
relates to existing data in our backend.
It will take an additional step to bring the newly created Fact in sync
with any preexisting categories or activities.
However, the decision to do it that way now allows for a clear
compartmentalization of those two independend steps.