Parametrized versions for existing objects.
We now provided heavily parametrized fixtures for our SQLAlchemy
objects. Actually using them in most of the tests blows them up quite a
bit (from <5s to >45s).
We will have to see if its worthwile, it shouldn't be realy needed but
its easy to reduce so we will roll with it for now.