Make category.name unique and refactor _add method.
A category.name should be unique. This is now stated in the BaseCategory
docstring as well as reflected by the sqlalchemy backend.
Our corresponding ``_add`` method now handles the expected error
accordingly.

One we turned the category_name into their PKs this should be cleaner.
Please note that ``save()`` does deliberatly not check if the category
to be saved's name is already present.