Remove legacy test.
Our parsing is realy just that: *parsing*. At that stage there is no
notion of "existsing" activity. All that happens is that we instanciate
a Fact with all the parsed information. This gets handed to the backend
which decides what to do / how to resolve any reference to existing
entities.