Made 'get_tmp_file_path' a BaseStore attribute.
Instead of beeing an independend helper, our tmpfile_path retrival
function is a method of ``BaseStore`` now. This should make the API more
natural.
We also removed the obsolete ``work_dir`` config dict key. Instead our
library will ony take fully qualified paths for now. That means that
instead of ``tmpfile_name`` there is not a ``tmpfile_path`` config dict
key. It is up to the client to provide the backend with proper paths.
In order to be more flexible in future about this it is advisable to
provide ``get_*_path`` hooks where possible instead of fetching the
config value directly. This way any additional future path-generation logic
does not trigger any changes with consuming code at all.

Closes: #97