Refactor store related config options.
Instead of passing a ``db_path`` config option and potentially parsing
it for all its semantical content we now accept a veriety of new config
optionss that each backend can retrieve as it sees fit. It is at eachs
backends discretion what keys it requires and to throw errors on invalid
configs. This should allow for better control and more better seperation
of concerns.
As far as our SQLAlchemy backend is conderned we provide a new store
method that will assemble a valid ``db_url`` to be passed on when
creating the engine or raise errors if required key/values are not
present in config.

Closes #107