Remove 'Fact.serialized_name' property and __repr__
As ``Fact.serialized_name`` is not realy needed in our rewrite we remove
it from the ``Fact`` class. This however meant that we had to transfer
some of its string construction logic to ``Fact.__str__``.
In order to avoid duplicating the effort with regards to
``Fact.__repr__`` and because issue #131 will take care of it in a
holistic way anyway we removed ``Fact.__repr__`` for now ass well.
Tests have been adjusted accordingly.

Note:
Our Fact class now provides a proper text representation even if it
misses any ``Fact.start`` value.

Closes: #146