Add custom __hash__ to objects(#137)
Under python3 any class with a custom ``__eq__`` method but no
``__hash__`` method is no longer ``hashable``. In order to avoid that we
add a custom ``__hash__`` method to ``Category``, ``Activity`` and
``Fact``. The method just computes the hash of ``self.as_tuple()`` which
should represent a 'serialized' version of out objects attributes.
A consequence is that *any* tuple that is ob the same nature will
match this hash.

Whilst I am not entirely sure that this is the best way to deal with
this, it does do for now.

Closes: #137