Add utilitly functions to handle config files
This commit adds various utility functions that provide consistent
behavior related to creating and retrieving config information.
As this means that our helper functions grow in scope, we also did use
the opportunity to split them up in a dedicated helper module.

In order to figure out the default location for config/log/etc
information within the filesystem we rely on the ``appdirs`` package.
For convenience we extended it so it makes sure the relevant directories
actually exist. In general this functionality may be usefull upstream
but so far we have had no feedback on this. If push comes to shove one
could also consider maintaining a fork if the project is dead.