Refactor FactManager._get_all(#138)
This commit removes ``FactManager._timeframe_is_free``. It's only
difference to ``FactManager._get_all()`` was that it would consider
facts that are only partially within the given timeframe. This behaviour
has been incorporated into ``_get_all()`` instead. By passing
``partial=True`` as a ``kwarg`` is the method will return all facts
within the timeframe that either start or end in it.
This can then be used to not only check if a timeframe is free but (in
case of an ``_update`` if the only fact blocking it is the one to be
updated.

We also used the oportunity to provide more and better tests.

Closes: #138