Add tags to ontology(#127)
This commit adds tags as optional information for facts to our ontology.
Facts can be used to add additional context to individual facts.

For that purpose a new ``objects.TAG`` has been added as well as a
corresponding SQLAlchemy class.

On the ``hamster`` level, ``Fact.tags`` is a set of ``objects.Tag``
instances. When a ``Fact`` is 'tupleized' this set is turned into a
``frozenset`` to be immutable.
On the SQLAlchemy level a intermediate table ``facttags`` has been
added.

Factories and fixtures have been adjusted accordingly to feature
radomized tags by default.

Because most sqlbackend related test do test that the db_instance
against the passed hamster instance it offen sufficed to make sure that
the used instance now features tags. For ``FactManager.__add``,`` __update``
and ``remove`` we added seperate tests to ensure and showcase tag related
beheviour.

Closes: #127