setup: Only require configparser when native version unavailable
On Python 3, configparser library is part of standard library. The configparser
package therefore is not needed. For this reason, distributions such as Arch Linux
do not even provide this package for Python 3, which in turn breaks the build.

This patch changes the setup.py to only add the configparser dependency when
native configparser is not available.