[LIB-254] time handling: Properly work with seconds
Previous versions of ``hamster-lib`` tried to be overly accomodating by
omitting seconds when parsing/displaying a times. This caused various
bugs and inconsistencies and is fixed by this commit.
While future versions may provide additional liberalizations our
solution now is the most accurate and strict one. It will be easier to
relax things on top of this than the other way round.

Please note that this also means that 'raw fact' parsing now requires
times to be specified as ``%H:%M:%S``