sqlalchemy: add method to check timeframe availability
This commit adds a new private method to the fact manager that handles
checking wether or not a the timeframe for a given fact is available.
This private method is also able to account for the passed fact being
the only fact taking up its own timeframe (relevant for update calls).
This allows for cleaner code in ``FactManager._add`` and
``FactManager._update``.

Future commits should rearange tests accordingly.

Ref: LIB-253