Explicitly pin 'factory-boy' version
So far, ``factory-boy`` was not part of our explicit requirements
because we relied on ``pytest-factoryboy`` to pull it as on of its
dependencies. This however prooved error prone as recent ``factory-boy``
versions raise failures (at least in our context).
In order to avoid this we move from installing ``factory-boy``
implicitly to making it part of our requirements as well as pin its last
known functioning version.

For details please also the [relevant ``pytest-factoryboy``
issue](https://github.com/pytest-dev/pytest-factoryboy/issues/47)